+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change Log
 
+## 2026-05-17
+- Global search (v0.0.23)
+
 ## 2026-05-16
-- Case-sensitive table name uniqueness; preserve case in storage codes (v0.0.22)
 - Datetime column type (v0.0.21)
 
 ## 2026-05-13